Improved views updating and checking

+++ lib/tasks/views_schema.rake
@@ -11,7 +11,7 @@ namespace :db do
           schema.puts ')'
         end
       end
-      File.delete('./db/views_schema.old')
+      File.delete('./db/views_schema.old') if File.exists?('./db/views_schema.old')
       File.rename('./db/views_schema.rb','./db/views_schema.old')
       File.rename('./db/views_schema.tmp','./db/views_schema.rb')
     end
